+++ tests/testthat/test-aggregates.R
@@ -403,3 +403,23 @@ test_that("region_aggregates() works as expected", {
   expect_equal(eur$S_units, amr$S_units)
   expect_equal(eur$r_EIOU, amr$r_EIOU)
 })
+
+
+test_that("despecified_aggregates() works as expected", {
+  mats_GBR <- UKEnergy2000mats %>%
+    tidyr::pivot_wider(names_from = matrix.name, values_from = matrix)
+
+  res <- mats_GBR %>%
+    despecified_aggregates()
+
+  # Check that S_units was aggregated
+  expect_true(all(res$S_units_aggregated[[1]] == 1))
+
+  # Check a few numbers
+  # Aggregated Crude
+  expect_equal(res$U_aggregated[[1]][1, 1], 48000)
+  # Aggregated Petrol
+  expect_equal(res$U_aggregated[[1]][5, 8], 27000)
+  # Aggregated Diesel
+  expect_equal(res$U_aggregated[[1]][2, 2], 15850)
+})