+++ NEWS.md
@@ -4,6 +4,8 @@ output: html_document
 ---
 
 
+* New function `despecified_aggregates()` that
+  aggregates to a piece of row and column labels. 
 * New argument `add_net_gross_cols` on `primary_aggregates()` that tells
   whether to add both net and gross primary energy columns.
   (Net and gross primary energy aggregate columns will contain identical values,